requirements.txt
requests
cacheout
azure-storage-blob==2.1.0
azure-storage-file
azure-storage-queue
azure-keyvault-secrets
azure-identity
azure-functions
azure-data-tables
marshmallow
python-json-logger
pytz
codetiming
aiohttp
azure-functions
azure-storage-queue
aiohttp
python-json-logger
pytz
azure-eventhub
azure-eventhub-checkpointstoreblob  
